added retries flag and unit test
